chore(pipeline) : Rename dags.dags in dags.dag_utils
And add an __init__.py at the upper level.

This way we can actually import modules correctly.

Before:

import dags.xxx would try to import a dag
import dags.virtualenvs would work but by chance because it was resolved
as a relative module
from dags import virtualenvs would fail cause "dags.dags" was not a module
and "dags" referred to the toplevel.
