+++ pipeline/dags/import_odspep.py
@@ -4,7 +4,7 @@ import airflow
 import pendulum
 from airflow.operators import empty, python
 
-from dags.virtualenvs import PYTHON_BIN_PATH
+from dag_utils.virtualenvs import PYTHON_BIN_PATH
 
 logger = logging.getLogger(__name__)
