+++ pipeline/dags/publish.py
@@ -2,7 +2,7 @@ import airflow
 import pendulum
 from airflow.operators import empty, python
 
-from dags.virtualenvs import PYTHON_BIN_PATH
+from dag_utils.virtualenvs import PYTHON_BIN_PATH
 
 default_args = {}
